requirements.txt
aiomysql==0.2.0
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
asttokens==2.4.1
async-timeout==4.0.3
blinker==1.7.0
certifi==2024.2.2
cffi==1.17.0
click==8.1.7
comm==0.2.2
cryptography==43.0.0
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
Flask==3.0.2
h11==0.14.0
hurry==1.1
hurry.filesize==0.9
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.23.0
iso8601==1.1.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
loguru==0.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==23.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.17.2
PyMySQL==1.1.1
pypika-tortoise==0.1.6
pytest==8.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pyzmq==26.0.0
redis==5.0.2
Redis-Sentinel-Url==1.0.1
rq==1.16.0
rq-dashboard==0.6.7.2
sentry-sdk==1.40.6
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
tomli==2.0.1
tornado==6.4
tortoise-orm==0.21.6
traitlets==5.14.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
wcwidth==0.2.13
Werkzeug==3.0.1
